Add Scapy packet sniffer with circular buffer
This commit adds a new module, oonib.sniffer, that contains the class
Sniffer. This class contains a thread that populates a circular buffer
with packets seen on an interface matching a filter (which currently is
set to only packets that are part of flows using the JSON HTTP headers
helper).

This methods for accessing this queue are thread-safe and allow for
helpers to query the queue for the raw packets sent and recieved during
measurements.

The sniffer is started in oonibackend.py when ooni-backend is started.